requirements.txt
flask
flask-socketio
mysql-connector-python
google-generativeai
python-dotenv